move approach discovery and strategyFactory to athena package
