+++ modules/text/module_text_llm/module_text_llm/__init__.py
@@ -1,3 +1,6 @@
 import dotenv
+from athena.approach_discovery.strategy_factory import SuggestionStrategyFactory
+from module_text_llm.approach_config import ApproachConfig
 
 dotenv.load_dotenv(override=True)
+strategyFactory = SuggestionStrategyFactory("module_text_llm", ApproachConfig)
